backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
icalendar==6.3.2
idna==3.11
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
